feat(deamon): implement deamon (v2)
This is a very large commit. It moves to the shell architecture outlined
in the included diagram (dod-shell.drawio).

The basic idea behind it is as follows: Rather than having each
component interface with the config files themselves and collect data on
their own they are only responsible for the UI portion of the component
offloading most other tasks to an always running deamon. In the future
this should also allow for easy IPC between the different components.